+++ src/components/05_pages/Donation/DonationSubmission/DonationSubmission.tsx
@@ -1,8 +1,8 @@
 // import { useId } from 'react';
-import DonationTable from './atoms/DonationTable';
-import DonationSubmission from './atoms/DonationSubmission';
+import DonationTable from '../../../03_organisms/Donaion/Donationtable/DonationTable';
+import DonationForm from '../../../03_organisms/Donaion/DonationForm/DonationForm';
 
-function DonationForm() {
+function DonationSubmission() {
   // const id = useId();
 
   return (
@@ -11,7 +11,7 @@ function DonationForm() {
         <h2 className="text-4xl">후원 신청</h2>
 
         <div className='max-w-[595px] flex flex-col gap-10 items-center'>
-          <DonationSubmission />
+          <DonationForm />
   
           <div className="w-full h-[1px] bg-gray-200"></div>
 
@@ -32,4 +32,4 @@ function DonationForm() {
   );
 }
 
-export default DonationForm;
+export default DonationSubmission;
