+++ src/components/05_pages/News/NewsDetails/NewsDetails.tsx
@@ -1,9 +1,9 @@
-import Title from "./atoms/Title";
+import NewsTitle from "../../../01_atoms/News/NewsTitle/NewsTitle";
 
 function NewsDetail() {
   return (
   <div className="max-w-[1440px] m-auto bg-white py-20">
-    <Title />
+    <NewsTitle />
     <div className="flex flex-col gap-3 items-center justify-center py-5">
       <p>[발행 날짜]</p>
       <p>기사 제목</p>
