--- src/pages/Donation/atoms/DonationFormInput.tsx
@@ -1,7 +1,0 @@
-function DonationFormInput () {
-  return (
-
-  )
-}
-
-export default DonationFormInput;
