--- src/pages/Donation/molecules/DonationButton.tsx
@@ -1,10 +1,0 @@
-function DonationButton() {
-  return (
-    <button 
-      className="font-bold text-blue-primary border-2 border-blue-primary rounded-[3px] py-2 px-8 hover:bg-blue-primary hover:text-white">
-      후원 기증 신청
-    </button>
-  )
-}
-
-export default DonationButton
