--- src/pages/Home/organisms/NewsCardList.tsx
@@ -1,8 +1,0 @@
-// type  NewsCardProps = {
-//   id: string;
-//   updated: string;
-//   title: string;
-//   type: string;
-//   content: string;
-// };
-
