+++ src/routes/navigation.tsx
@@ -29,7 +29,7 @@ const navigationItems = [
     path: '/News',
     text: '뉴스',
     lazy: async () => {
-      const Module = await import('@/pages/News/News');
+      const Module = await import('@/components/05_pages/News/News/News');
       return { Component: Module.default };
     },
   },
@@ -38,7 +38,7 @@ const navigationItems = [
     path: '/NewsDetails',
     text: '뉴스 내용',
     lazy: async () => {
-      const Module = await import('@/pages/News/NewsDetails');
+      const Module = await import('@/components/05_pages/News/NewsDetails/NewsDetails');
       return { Component: Module.default };
     },
   },
@@ -155,7 +155,7 @@ const navigationItems = [
     path: '/Donation',
     text: '기부페이지',
     lazy: async () => {
-      const Module = await import('@/pages/Donation/Donation');
+      const Module = await import('@/components/05_pages/Donation/Donation/Donation');
       return { Component: Module.default };
     },
   },
@@ -164,7 +164,7 @@ const navigationItems = [
     path: '/DonationForm',
     text: '기부페이지',
     lazy: async () => {
-      const Module = await import('@/pages/Donation/DonationForm');
+      const Module = await import('@/components/05_pages/Donation/DonationSubmission/DonationSubmission');
       return { Component: Module.default };
     },
   },
